requirements.txt
streamlit==1.32.0
pandas==2.2.1
numpy==1.26.4
python-dotenv==1.0.1
requests==2.31.0
groq
openpyxl>=3.1.0
reportlab>=4.0.0
